+++ chaostoolkit/cli.py
@@ -1,7 +1,7 @@
 import json
 import os
 import uuid
-from typing import Any, Dict, List
+from typing import Any, Dict, List, Optional
 
 import click
 from chaoslib import __version__ as chaoslib_version
@@ -26,7 +26,7 @@ from chaoslib.settings import (
     locate_settings_entry,
     save_settings,
 )
-from chaoslib.types import Activity, Discovery, Experiment, Journal, Schedule
+from chaoslib.types import Activity, Discovery, Dry, Experiment, Journal, Schedule
 from click_plugins import with_plugins
 
 try:
@@ -135,7 +135,10 @@ def validate_vars(
     help="Path where to save the journal from the execution.",
 )
 @click.option(
-    "--dry", is_flag=True, help="Run the experiment without executing activities."
+    "--dry",
+    type=click.Choice(["probes", "actions", "activities", "pause"]),
+    show_default=False,
+    help="Run the experiment without executing the chosen strategy.",
 )
 @click.option(
     "--no-validation",
@@ -211,7 +214,7 @@ def run(
     ctx: click.Context,
     source: str,
     journal_path: str = "./journal.json",
-    dry: bool = False,
+    dry: Optional[str] = None,
     no_validation: bool = False,
     no_exit: bool = False,
     no_verify_tls: bool = False,
@@ -249,7 +252,7 @@ def run(
             logger.debug(x)
             ctx.exit(1)
 
-    experiment["dry"] = dry
+    experiment["dry"] = Dry.from_string(dry)
     settings.setdefault("runtime", {}).setdefault("rollbacks", {}).setdefault(
         "strategy", rollback_strategy
     )
@@ -267,7 +270,8 @@ def run(
     )
     has_deviated = journal.get("deviated", False)
     has_failed = journal["status"] != "completed"
-
+    if "dry" in journal["experiment"]:
+        journal["experiment"]["dry"] = dry
     with open(journal_path, "w") as r:
         json.dump(journal, r, indent=2, ensure_ascii=False, default=encoder)